Refactor files to comply with Ruff
