+++ kwave/kWaveSimulation_helper/display_simulation_params.py
@@ -1,6 +1,7 @@
 import numpy as np
 
-from kwave import kWaveGrid, kWaveMedium
+from kwave.kgrid import kWaveGrid
+from kwave.kmedium import kWaveMedium
 from kwave.utils.data import scale_SI
