+++ kwave/reconstruction/converter.py
@@ -4,7 +4,7 @@ import numpy as np
 import uff
 from uff.linear_array import LinearArray
 
-from kwave import NotATransducer
+from kwave.ktransducer import NotATransducer
 
 
 def build_channel_data(sensor_data: np.ndarray,
