+++ pyproject.toml
@@ -81,3 +81,4 @@ line-length = 200
 "kwave/ksource.py" = ["F821"]
 # create_storage_variables.py references a few functions that are not ported yet.
 "kwave/kWaveSimulation_helper/create_storage_variables.py" = ["F821"]
+"kwave/kWaveSimulation_helper/__init__.py" = ["F401"]
